refactor(toaster): separate utils
