+++ src/features/toaster/components/toaster.tsx
@@ -6,18 +6,17 @@ import { toast } from "sonner";
 
 import { Toaster as ShadcnToaster } from "@/components/ui/sonner";
 
-import { useSavedToast } from "../hooks/use-saved-toast";
+import { deleteSavedToast } from "../utils/delete-saved-toast";
+import { parseSavedToast } from "../utils/parse-saved-toast";
 
 export function Toaster(props: ComponentProps<typeof ShadcnToaster>) {
-  const savedToastReturn = useSavedToast();
-
   useEffect(() => {
-    const { savedToast, deleteSavedToast } = savedToastReturn;
+    const savedToast = parseSavedToast();
     if (savedToast == null) {
       return;
     }
     toast[savedToast.level](savedToast.message);
     deleteSavedToast();
-  }, [savedToastReturn]);
+  }, []);
   return <ShadcnToaster {...props} />;
 }
